scraper/requirements.txt
instagrapi==2.3.0
supabase==2.15.3
python-dotenv==1.0.1
